rebasing from main

+++ src/lib/onesource.cpp
@@ -816,35 +816,34 @@ void onesource::generatePDF(vector<SED*> &fulllib, const vector<size_t> &va, con
   // need to convert into 1 dimension array for openMP reduction
   // 0:["MASS"] / 1:["SFR"] / 2:["SSFR"] / 3:["LDUST"] / 4:["LIR"] / 5:["AGE"] / 6:["COL1"] / 7:["COL2"] / 8:["MREF"]/ 9:["MIN_ZG"] / 10:["MIN_ZQ"] / 11:["BAY_ZG"] / 12:["BAY_ZQ"]
   double PDFzloc[pdfmap[11].size()];
-  for (int i=0; i<pdfmap[11].size(); ++i) PDFzloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[11].size(); ++i) PDFzloc[i] = 0.0;
 
   double PDFzqloc[pdfmap[12].size()];
-  for (int i=0; i<pdfmap[12].size(); ++i) PDFzqloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[12].size(); ++i) PDFzqloc[i] = 0.0;
 
   double PDFmassloc[pdfmap[0].size()];
-  for (int i=0; i<pdfmap[0].size(); ++i) PDFmassloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[0].size(); ++i) PDFmassloc[i] = 0.0;
 
   double PDFSFRloc[pdfmap[1].size()];
-  for (int i=0; i<pdfmap[1].size(); ++i) PDFSFRloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[1].size(); ++i) PDFSFRloc[i] = 0.0;
   
   double PDFsSFRloc[pdfmap[2].size()];
-  for (int i=0; i<pdfmap[2].size(); ++i) PDFsSFRloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[2].size(); ++i) PDFsSFRloc[i] = 0.0;
   
   double PDFAgeloc[pdfmap[5].size()];
-  for (int i=0; i<pdfmap[5].size(); ++i) PDFAgeloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[5].size(); ++i) PDFAgeloc[i] = 0.0;
   
   double PDFLdustloc[pdfmap[3].size()];
-  for (int i=0; i<pdfmap[3].size(); ++i) PDFLdustloc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[3].size(); ++i) PDFLdustloc[i] = 0.0;
   
   double PDFcol1loc[ pdfmap[6].size() ];
-  for (int i=0; i<pdfmap[6].size(); ++i) PDFcol1loc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[6].size(); ++i) PDFcol1loc[i] = 0.0;
   
   double PDFcol2loc[ pdfmap[7].size() ];
-  for (int i=0; i<pdfmap[7].size(); ++i) PDFcol2loc[i] = 0.0;
+  for (size_t i=0; i<pdfmap[7].size(); ++i) PDFcol2loc[i] = 0.0;
   
   double PDFmrefloc[ pdfmap[8].size() ];
-  for (int i=0; i<pdfmap[8].size(); ++i) PDFmrefloc[i] = 0.0;
-
+  for (size_t i=0; i<pdfmap[8].size(); ++i) PDFmrefloc[i] = 0.0;
 
   // Decide if the uncertainties on the rest-frame colors should be analysed
   bool colAnalysis;
@@ -1031,7 +1030,7 @@ void onesource::generatePDF_IR(vector<SED*>& fulllib){
 
   //4:["LIR"] 
   double PDFlirloc[pdfmap[4].size()];
-  for (int i = 0; i < pdfmap[4].size(); ++i) PDFlirloc[i] = 0.0;
+  for (size_t i = 0; i < pdfmap[4].size(); ++i) PDFlirloc[i] = 0.0;
 
   // parrallellize over each SED
   #ifdef _OPENMP
